refactor: move processNerfs to its own file
